rename: spec 구조 수정
